pyproject.toml
[project]
name = "social-support-agentic-ai"
version = "0.1.0"
description = ""
authors = [
    {name = "Md Marghub Akhtar",email = "marghub79@gmail.com"}
]
readme = "README.md"

[tool.poetry.dependencies]

python = ">=3.11,<3.12"

# Web Frameworks
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
python-multipart = "^0.0.6"

# Streamlit UI
streamlit = "^1.31.0"

# LLM & Agent Framework (Assignment Required)

langchain-core = "0.2.41"
langchain-community = "0.2.11"
langgraph = "0.2.36"
langchain-ollama = "0.1.3"  # For Mistral integration

# Ollama Client (for your Mistral model)
ollama = ">=0.3.0,<1.0.0"

# Cohere (for your embeddings)
cohere = "^4.47"

# Langfuse (Observability)
langfuse = "^2.20.0"

# ML/AI Models 
scikit-learn = "^1.4.0"
xgboost = "^2.0.3"
pandas = "^2.2.0"
numpy = ">=1.26.3,<2.0.0"
scipy = "^1.12.0"
shap = ">=0.44.0,<0.50.0"  # For explainability (compatible with NumPy 1.26)
imbalanced-learn = "^0.12.0"

# Document Processing (Multi-modal)

PyPDF2 = "^3.0.1"
pdfplumber = "^0.10.3"
python-docx = "^1.1.0"
openpyxl = "^3.1.2"
pillow = "^10.2.0"


# Lightweight Databases (M1 Optimized)
sqlalchemy = "^2.0.25"
tinydb = "^4.8.0"
chromadb = "^0.4.15"  # Lightweight vector DB , stable for M1
networkx = "^3.2.1"  # Graph processing (instead of Neo4j)

# Utilities
python-dotenv = "^1.0.1"
python-dateutil = "^2.8.2"
pytz = "^2024.1"
tenacity = "^8.2.3"
aiofiles = "^23.2.1"
psutil = "^5.9.8"
httpx = ">=0.27,<0.29"

# Security
cryptography = "^42.0.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}

# Logging
loguru = "^0.7.2"
plotly = "^6.5.0"
pymongo = "^4.15.5"

[tool.poetry.group.multimodal.dependencies]
opencv-python = "^4.9.0"
pytesseract = "^0.3.10"
pdf2image = "^1.17.0"

[tool.poetry.group.dev.dependencies]
# Development tools
pytest = "^8.0.0"
pytest-asyncio = "^0.23.4"
pytest-cov = "^4.1.0"
black = "^24.1.1"
flake8 = "^7.0.0"
mypy = "^1.8.0"
pre-commit = "^3.6.0"
ipython = "^8.20.0"
jupyter = "^1.0.0"

# Testing
faker = "^22.6.0"

[tool.poetry]
package-mode = false

[tool.poetry.group.docs.dependencies]
# Documentation
mkdocs = "^1.5.3"
mkdocs-material = "^9.5.3"

[tool.poetry.scripts]
# Custom commands
start-api = "uvicorn src.api.main:app --reload --port 8000"
start-ui = "streamlit run streamlit_app/app.py --server.port 8501"
train-models = "python scripts/train_all_models.py"
test-setup = "python scripts/test_setup.py"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --cov=src --cov-report=html --cov-report=term"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
